feat: restructured code and removed console logs
